:recycle: Refactor React components for separation of concerns and scalability
