+++ src/App.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import Quiz from './components/Quiz/Quiz'
+import Quiz from './components/Quiz/QuizCard'
 
 export default function App (){
+
     return ( <Quiz />)
 }
